+++ templates/editor.php
@@ -34,6 +34,7 @@
         data-directtoken="<?php p($_["directToken"]) ?>"
         data-version="<?php p($_["version"]) ?>"
         data-template="<?php p($_["isTemplate"]) ?>"
+        data-anchor="<?php p($_["anchor"]) ?>"
         data-inframe="<?php p($_["inframe"]) ?>"></div>
 
     <?php if (!empty($_["documentServerUrl"])) { ?>